Move get_cached_with_mtime to separate module
